Revert "remove Authentication.ts -> Delegate.ts require cycle"
This reverts commit 06939df6affb10891f317b27b8560f2de0ffaaf1.
