+++ src/libs/actions/Delegate.ts
@@ -10,11 +10,10 @@ import * as SequentialQueue from '@libs/Network/SequentialQueue';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import type {Delegate, DelegatedAccess, DelegateRole} from '@src/types/onyx/Account';
+import type Response from '@src/types/onyx/Response';
 import {confirmReadyToOpenApp, openApp} from './App';
-import {KEYS_TO_PRESERVE_DELEGATE_ACCESS} from './const';
-import isConnectedAsDelegate from './isConnectedAsDelegate';
-import restoreDelegateSession from './restoreDelegateSession';
 import updateSessionAuthTokens from './Session/updateSessionAuthTokens';
+import updateSessionUser from './Session/updateSessionUser';
 
 let delegatedAccess: DelegatedAccess;
 Onyx.connect({
@@ -23,6 +22,19 @@ Onyx.connect({
         delegatedAccess = val?.delegatedAccess ?? {};
     },
 });
+
+const KEYS_TO_PRESERVE_DELEGATE_ACCESS = [
+    ONYXKEYS.NVP_TRY_FOCUS_MODE,
+    ONYXKEYS.PREFERRED_THEME,
+    ONYXKEYS.NVP_PREFERRED_LOCALE,
+    ONYXKEYS.SESSION,
+    ONYXKEYS.IS_LOADING_APP,
+    ONYXKEYS.CREDENTIALS,
+
+    // We need to preserve the sidebar loaded state since we never unrender the sidebar when connecting as a delegate
+    // This allows the report screen to load correctly when the delegate token expires and the delegate is returned to their original account.
+    ONYXKEYS.IS_SIDEBAR_LOADED,
+];
 
 function connect(email: string) {
     if (!delegatedAccess?.delegators) {
@@ -383,6 +395,10 @@ function clearAddDelegateErrors(email: string, fieldName: string) {
     });
 }
 
+function isConnectedAsDelegate() {
+    return !!delegatedAccess?.delegate;
+}
+
 function removePendingDelegate(email: string) {
     if (!delegatedAccess?.delegates) {
         return;
@@ -531,6 +547,19 @@ function clearDelegateRolePendingAction(email: string) {
     ];
 
     Onyx.update(optimisticData);
+}
+
+function restoreDelegateSession(authenticateResponse: Response) {
+    Onyx.clear(KEYS_TO_PRESERVE_DELEGATE_ACCESS).then(() => {
+        updateSessionAuthTokens(authenticateResponse?.authToken, authenticateResponse?.encryptedAuthToken);
+        updateSessionUser(authenticateResponse?.accountID, authenticateResponse?.email);
+
+        NetworkStore.setAuthToken(authenticateResponse.authToken ?? null);
+        NetworkStore.setIsAuthenticating(false);
+
+        confirmReadyToOpenApp();
+        openApp();
+    });
 }
 
 export {
